no pid exception

+++ cerise.py
@@ -119,11 +119,13 @@ def account():
 @login_required
 def project():
     form = ProjectForm()
-    try:
-        os.kill(current_user.pid, 0)
-        processLive = True
-    except OSError:
-        processLive = False
+    processLive = False
+    if current_user.pid:
+        try:
+            os.kill(current_user.pid, 0)
+            processLive = True
+        except OSError:
+            processLive = False
     if request.method == 'GET':
         currentProject = request.args.get('name')
         project = current_user.projects.get(name=currentProject)
